Add a path option to "gr tag discover"
If a user has a particularly complex home directory or
would like the convenience of discovering only some subsection
of git repositories, it may be useful to specify a specific path
to discover within.

This commit provides this fucntionality by adding an optional <path>
argument to the tag discover command. When <path> is omitted the
discover path will default to the home directory to maintain existing
functionality. Otherwise the path will be used as the root of the
discovery search.

This fixes issue #7